move some classes to subfolders or inner classes
