--- Src/MoneyFox.Core/Model/StatisticSelectorType.cs
@@ -1,25 +1,0 @@
-﻿using MoneyManager.Foundation;
-
-namespace MoneyFox.Core.Model
-{
-    /// <summary>
-    ///     Represents a item for the selector to choose the statistic.
-    /// </summary>
-    public class StatisticSelectorType
-    {
-        /// <summary>
-        ///     Name of the statistic
-        /// </summary>
-        public string Name { get; set; }
-
-        /// <summary>
-        ///     Short description for the statistic
-        /// </summary>
-        public string Description { get; set; }
-
-        /// <summary>
-        ///     Type of this item.
-        /// </summary>
-        public StatisticType Type { get; set; }
-    }
-}
