+++ Src/MoneyFox.Core/Statistics/DataProviders/CashFlowDataProvider.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using MoneyFox.Core.Interfaces;
 using MoneyFox.Core.Model;
+using MoneyFox.Core.Statistics.Models;
 using MoneyFox.Foundation.Model;
 using MoneyFox.Foundation.Resources;
 using MoneyManager.Foundation;
 
-namespace MoneyManager.Core.StatisticDataProvider
+namespace MoneyFox.Core.Statistics.DataProviders
 {
     public class CashFlowDataProvider : IStatisticProvider<CashFlow>
     {
