+++ Src/MoneyFox.Core/Statistics/DataProviders/CategorySpreadingDataProvider.cs
@@ -3,10 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using MoneyFox.Core.Interfaces;
 using MoneyFox.Core.Model;
+using MoneyFox.Core.Statistics.Models;
 using MoneyFox.Foundation.Model;
 using MoneyManager.Foundation;
 
-namespace MoneyManager.Core.StatisticDataProvider
+namespace MoneyFox.Core.Statistics.DataProviders
 {
     public class CategorySpreadingDataProvider : IStatisticProvider<IEnumerable<StatisticItem>>
     {
