+++ Src/MoneyFox.Core/Statistics/DataProviders/CategorySummaryDataProvider.cs
@@ -4,10 +4,10 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using MoneyFox.Core.Interfaces;
 using MoneyFox.Core.Model;
-using MoneyFox.Foundation.Model;
+using MoneyFox.Core.Statistics.Models;
 using MoneyManager.Foundation;
 
-namespace MoneyManager.Core.StatisticDataProvider
+namespace MoneyFox.Core.Statistics.DataProviders
 {
     public class CategorySummaryDataProvider : IStatisticProvider<IEnumerable<StatisticItem>>
     {
