+++ Src/MoneyFox.Core/ViewModels/StatisticCategorySpreadingViewModel.cs
@@ -3,8 +3,9 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using MoneyFox.Core.Interfaces;
 using MoneyFox.Core.Model;
+using MoneyFox.Core.Statistics.DataProviders;
+using MoneyFox.Core.Statistics.Models;
 using MoneyFox.Foundation.Model;
-using MoneyManager.Core.StatisticDataProvider;
 using OxyPlot;
 using OxyPlot.Series;
 using PropertyChanged;
