+++ Src/MoneyFox.Core/ViewModels/StatisticCategorySummaryViewModel.cs
@@ -1,9 +1,10 @@
 ﻿using System.Collections.ObjectModel;
 using MoneyFox.Core.Interfaces;
 using MoneyFox.Core.Model;
+using MoneyFox.Core.Statistics.DataProviders;
+using MoneyFox.Core.Statistics.Models;
 using MoneyFox.Core.ViewModels;
 using MoneyFox.Foundation.Model;
-using MoneyManager.Core.StatisticDataProvider;
 using PropertyChanged;
 
 namespace MoneyManager.Core.ViewModels
