+++ Src/MoneyFox.Core/ViewModels/StatisticMonthlyExpensesViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using MoneyFox.Core.Interfaces;
+using MoneyFox.Core.Statistics.DataProviders;
 using MoneyFox.Core.ViewModels;
-using MoneyManager.Core.StatisticDataProvider;
 using OxyPlot;
 using OxyPlot.Axes;
 using OxyPlot.Series;
