+++ Src/MoneyFox.Core/ViewModels/StatisticSelectorViewModel.cs
@@ -3,8 +3,6 @@ using GalaSoft.MvvmLight;
 using GalaSoft.MvvmLight.Command;
 using GalaSoft.MvvmLight.Views;
 using MoneyFox.Core.Constants;
-using MoneyFox.Core.Model;
-using MoneyFox.Foundation.Model;
 using MoneyFox.Foundation.Resources;
 using MoneyManager.Foundation;
 
@@ -77,5 +75,26 @@ namespace MoneyFox.Core.ViewModels
                     break;
             }
         }
+    }
+
+    /// <summary>
+    ///     Represents a item for the selector to choose the statistic.
+    /// </summary>
+    public class StatisticSelectorType
+    {
+        /// <summary>
+        ///     Name of the statistic
+        /// </summary>
+        public string Name { get; set; }
+
+        /// <summary>
+        ///     Short description for the statistic
+        /// </summary>
+        public string Description { get; set; }
+
+        /// <summary>
+        ///     Type of this item.
+        /// </summary>
+        public StatisticType Type { get; set; }
     }
 }
