+++ Src/MoneyFox.Windows/Views/StatisticSelectorView.xaml
@@ -6,11 +6,9 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:core="using:Microsoft.Xaml.Interactions.Core"
-    xmlns:model1="using:MoneyFox.Foundation.Model"
     xmlns:converter1="using:MoneyFox.Windows.Converter"
     xmlns:userControls1="using:MoneyFox.Windows.Views.UserControls"
     xmlns:viewModels="using:MoneyFox.Core.ViewModels"
-    xmlns:model="using:MoneyFox.Core.Model"
     d:DataContext="{d:DesignInstance viewModels:StatisticSelectorViewModel, IsDesignTimeCreatable=True}"
     mc:Ignorable="d">
 
@@ -18,7 +16,7 @@
         <converter1:SelectedConverter x:Key="SelectedConverter" />
 
         <DataTemplate x:Key="StatistcTypeTemplate"
-                      x:DataType="model:StatisticSelectorType">
+                      x:DataType="viewModels:StatisticSelectorType">
             <StackPanel HorizontalAlignment="Stretch"
                         Background="Transparent"
                         Margin="0,2,0,4">
