+++ Src/MoneyManager.sln
@@ -16,8 +16,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyManager.Core.Tests", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyManager.Windows.Tests", "MoneyManager.Windows.Tests\MoneyManager.Windows.Tests.csproj", "{E70B2E42-3944-4D29-AE11-8D7104FBB3A4}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tasks", "Tasks", "{9705BAF6-DD11-40F6-B338-E6929B948FF2}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyManager.Windows.DataAccess.Tests", "MoneyManager.Windows.DataAccess.Tests\MoneyManager.Windows.DataAccess.Tests.csproj", "{09B31D96-0408-4197-A7AC-3760D757BE9D}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyFox.Core", "MoneyFox.Core\MoneyFox.Core.csproj", "{2F774D5A-E9B5-4AB2-84AA-51A6E9BDB456}"
@@ -445,7 +443,6 @@ Global
 		{8D8635AE-2730-405B-B932-54F39A7FE2F5} = {D6E6E944-C844-4882-A72A-F86C17196117}
 		{E70B2E42-3944-4D29-AE11-8D7104FBB3A4} = {D6E6E944-C844-4882-A72A-F86C17196117}
 		{09B31D96-0408-4197-A7AC-3760D757BE9D} = {D6E6E944-C844-4882-A72A-F86C17196117}
-		{55DD9633-DB49-4E60-8805-4DB14603B160} = {9705BAF6-DD11-40F6-B338-E6929B948FF2}
 	EndGlobalSection
 	GlobalSection(MonoDevelopProperties) = preSolution
 		version = 0.12.0.0
